requirements.txt
## Dependencies
pyusb
pyvisa
pyvisa-py
pyserial